feat: isolated the network layer creation logic
