+++ lib/src/functions/api/types.js
@@ -3,6 +3,11 @@ const { promisify } = require('util')
 const exec = promisify(require('child_process').exec)
 const writeFile = require('../../utils/writeFile')
 
+/**
+ * @param {Object} args
+ * @param {Boolean} args.express
+ * @param {String} args.projectName
+ */
 module.exports = async ({ express, projectName }) => {
   const createFoldersCommand = `mkdir ${projectName}/src/@types \
 ${projectName}/src/@types/models \
